src/requirements.txt
PyYAML==5.1
requests==2.22.0
